Move config manager to service directory
