+++ src/core/migration.mjs
@@ -1,4 +1,4 @@
-import ConfigManager from "/core/helpers/config-manager.mjs";
+import ConfigManager from "/core/services/config-manager.mjs";
 
 browser.runtime.onInstalled.addListener(async (details) => {
