+++ src/core/services/config-manager.mjs
