+++ src/views/options/main.mjs
@@ -1,6 +1,6 @@
 import { fetchHTMLAsFragment } from "/core/utils/commons.mjs";
 
-import ConfigManager from "/core/helpers/config-manager.mjs";
+import ConfigManager from "/core/services/config-manager.mjs";
 
 import DefaultConfig from "/resources/configs/defaults.mjs";
